% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailypotus.R
\name{dailypotus}
\alias{dailypotus}
\alias{generic_timeline}
\alias{daily_wiki_trump}
\alias{daily_wiki_biden}
\title{Get presidential timelines from wikipedia}
\usage{
generic_timeline(qs, url, pres)

daily_wiki_trump()

daily_wiki_biden()
}
\description{
Get presidential timelines from wikipedia
}
